+++ src/main/java/com/zaia08/hub_estudos/repositories/CursoRepository.java
@@ -4,6 +4,6 @@ import com.zaia08.hub_estudos.Model.Curso;
 import jakarta.persistence.Id;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CursoRepository extends JpaRepository<Curso, Integer> {
+public interface CursoRepository extends JpaRepository<Curso, Long> {
 
 }
